Move uniform out of extract_component
